actually done

+++ texture.hpp
@@ -18,7 +18,7 @@ class aTexture
 		bool loadFromFile(std::string path);
 		bool loadFromRenderedText(std::string textureText,SDL_Color textColor);
 		void free();
-		void render(int x,int y,SDL_Rect* clip = NULL,double angle = 0.0,SDL_Point* center = NULL,SDL_RendererFlip flip = SDL_FLIP_NONE);
+		void render(int x,int y,SDL_Rect* clip = NULL,int angle = 0,SDL_Point* center = NULL,SDL_RendererFlip flip = SDL_FLIP_NONE);
 		int gW();
 		int gH();
 	private:
